refactor(claude): colocate inventory binding helpers
