+++ crates/ctx-history-provider-claude-cursor/src/claude/nativepath/source_backed/binding.rs
@@ -1,4 +1,8 @@
 use super::*;
+use ctx_history_source_io::{
+    NON_REGULAR_PROVIDER_SOURCE_REASON, REPARSE_PROVIDER_SOURCE_REASON,
+    SYMLINK_PROVIDER_SOURCE_REASON,
+};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
@@ -26,4 +30,42 @@ pub(super) fn relative_to_authority(
             path: path.to_path_buf(),
             reason: "Claude transcripts must remain below their selected authority",
         })
+}
+
+pub(super) fn contract(error: impl std::fmt::Display) -> CaptureError {
+    CaptureError::InvalidPayload(error.to_string())
+}
+
+pub(super) fn is_quarantinable_claude_leaf_error(error: &CaptureError) -> bool {
+    matches!(error, CaptureError::Io(source) if source.kind() == io::ErrorKind::PermissionDenied)
+        || matches!(
+            error,
+            CaptureError::InvalidProviderTranscriptPath { reason, .. }
+                if *reason == SYMLINK_PROVIDER_SOURCE_REASON
+                    || *reason == REPARSE_PROVIDER_SOURCE_REASON
+                    || *reason == NON_REGULAR_PROVIDER_SOURCE_REASON
+        )
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(super) enum ClaudeSourceClaim {
+    New,
+    Duplicate,
+}
+
+pub(super) fn claim_claude_source(
+    claimed: &mut HashMap<[u8; 32], SourceKey>,
+    source: &SourceKey,
+) -> Result<ClaudeSourceClaim> {
+    let digest = source.exact_descriptor_digest();
+    if let Some(previous) = claimed.get(&digest) {
+        if previous.exact_descriptor_eq(source) {
+            return Ok(ClaudeSourceClaim::Duplicate);
+        }
+        return Err(CaptureError::InvalidPayload(
+            "Claude source descriptor digest collision".to_owned(),
+        ));
+    }
+    claimed.insert(digest, source.clone());
+    Ok(ClaudeSourceClaim::New)
 }
